Update crafting.lua with all legendary recipe items
I updated the crafting.lua file with every available legendary recipe in the game. 
At least the ones that actually work. I named the iconic items, since that might be useful for someone?

I made small categories throughout the code, should more recipes/items be added later into the game. Figured it might be easier with some categories.

Note, this is only the legendary version of the recipes. There are still a lot more recipes in the game. But since the current command in the console is for legendary recipes, I figured I just wanted to add those. 

I hope I didn't mess anything up. Just used the in-house Github editor.