Updated crafting.lua with new commands + new items
1) 
Added three new commands:
    Crafting.AddModRecipes()
    Crafting.AddMedRecipes()
    Crafting.AddGrenadeRecipes()

and integrated them to the
Crafting.AddAll()

2) 
I forgot the fact that in order to craft some of the legendary iconic weapons and quick-hacks, you'll need the recipe for the common and rare versions too.
Hopefully this little addition should fix it. 
I basically just added the other recipes to the individual iconic weapons and quickhacks that had other recipes available beyond the one(s) tagged legendary. I also added all the quickhacks, even the ones not available as legendary (some of them only goes to "Lvl3" which is "epic" quality ingame. Hopefully this wont be an issue.

Note:
I'm not sure how the crafting of some of the iconic weapons work. Since you would probably need 1 "rare" or "uncommon" version of the weapon, even in order to craft the epic version of it. You're usually given that, by the end of the quest where you're originally supposed to get it from (as a reward). And then you can use the recipes in order to upgrade it.

So I'm not sure if adding the recipes themselves would be enough. You might need to add the "base" (the rewarded) version of the weapon, in order to be able to use the iconic weapon recipes.. Anyways, that'll be for a later fix / version I guess.