Added facts.lua
Added a few quest trigger protocols, making the player able to change the outcome of an already completed quest.

Some of these are pretty "spoilery", if you haven't played the game, so I'm not sure what to do about that.

Also, not so sure about how the romance options work. Apparently you have to apply them, each time you reload a savegame.