Fixed legendary item upgrading
Fixes this issue:
https://github.com/WolvenKit/BraindanceProtocol/issues/51

- Adds a few protocols, changes the old one:
	Legend.UpgradeWeapons()
	Legend.UpgradeArmor()
	Legend.UpgradeMods()
	Legend.UpgradeCW()

The old Legend.Create() is now Legend.UpgradeAll() and includes all of the commands above.

Not sure if we should include the CW upgrade or not, since it's purely for aesthetics. It doesn't really upgrade the item.

Oh, and I also changed a bit in the protocols. Added a new category, "item modifications", instead of putting these under cheats. Seemed more appropriate.